pyproject.toml
[tool.poetry]
name = "langchain-study"
version = "0.1.0"
description = ""
authors = ["user00"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "1.40.6"
langchain-core = "0.2.30"
langchain-openai = "0.1.21"
langchain-community = "0.2.12"
langchain-text-splitters = "0.2.2"
langchain-chroma = "0.1.2"
streamlit = "1.38.0"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
